:truck: refactor as multiple packages
