+++ quant_core/pyproject.toml
@@ -1,0 +1,19 @@
+[tool.poetry]
+name = "quant-core"
+version = "0.1.0"
+description = "Core ML prediction logic for quant_tick and quant_horizon"
+authors = ["Alex <globophobe@gmail.com>"]
+readme = "README.md"
+license = "MIT"
+packages = [{ include = "quant_core" }]
+
+[tool.poetry.dependencies]
+python = ">=3.10"
+lightgbm = "*"
+scikit-learn = "*"
+pandas = "*"
+numpy = "*"
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
