+++ quant_core/quant_core/__init__.py
@@ -1,0 +1,3 @@
+"""quant_core - Pure Python ML prediction logic."""
+
+__version__ = "0.1.0"
