+++ quant_core/quant_core/constants.py
@@ -1,0 +1,8 @@
+"""Constants for ML prediction."""
+
+# Sentinel value for missing features
+MISSING_SENTINEL = -999.0
+
+# Default bounds grid for training and inference
+DEFAULT_WIDTHS = [0.02, 0.03, 0.04, 0.05, 0.07, 0.10]
+DEFAULT_ASYMMETRIES = [-0.4, -0.2, 0.0, 0.2, 0.4]
