+++ quant_horizon/.dockerignore
@@ -1,0 +1,21 @@
+__pycache__
+*.pyc
+*.pyo
+*.pyd
+.Python
+*.so
+*.egg
+*.egg-info
+dist
+build
+.git
+.gitignore
+.env
+.venv
+venv/
+ENV/
+.pytest_cache
+.coverage
+htmlcov/
+*.log
+.DS_Store
