+++ quant_horizon/pyproject.toml
@@ -1,0 +1,20 @@
+[tool.poetry]
+name = "quant-horizon"
+version = "0.1.0"
+description = "FastAPI prediction service for quant_tick ML models"
+authors = ["Alex <globophobe@gmail.com>"]
+readme = "README.md"
+license = "MIT"
+packages = [{ include = "quant_horizon" }]
+
+[tool.poetry.dependencies]
+python = ">=3.10"
+fastapi = "*"
+uvicorn = {extras = ["standard"], version = "*"}
+quant-core = {path = "../quant_core", develop = true}
+google-cloud-storage = "*"
+pydantic = "*"
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
