+++ quant_horizon/quant_horizon/__init__.py
@@ -1,0 +1,3 @@
+"""quant_horizon - FastAPI prediction service."""
+
+__version__ = "0.1.0"
